+++ src/main/java/com/ar/routes/domain/repository/RouteRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RouteRepository extends JpaRepository<Route, Integer> {
+public interface RouteRepository extends JpaRepository<Route, Long> {
     List<Route> findByOrigin(Station origin);
 }
