+++ src/main/java/com/ar/routes/domain/service/StationService.java
@@ -1,6 +1,7 @@
 package com.ar.routes.domain.service;
 
 import com.ar.routes.domain.model.Station;
+import com.ar.routes.domain.model.dto.CreateEditStationDto;
 import org.apache.coyote.BadRequestException;
 
 import java.util.List;
@@ -8,5 +9,5 @@ import java.util.List;
 public interface StationService {
     Station add(String station);
     List<Station> getAll();
-    void delete(Integer id) throws BadRequestException;
+    void delete(Long id) throws BadRequestException;
 }
